moved static data to src/data
